feat(gpu): signed scalar rotate
