+++ tfhe/src/high_level_api/integers/unsigned/scalar_ops.rs
@@ -770,7 +770,7 @@ generic_integer_impl_scalar_operation!(
                 InternalServerKey::Cuda(cuda_key) => {
                     let inner_result = with_thread_local_cuda_stream(|stream| {
                         cuda_key.key.scalar_rotate_left(
-                            &lhs.ciphertext.on_gpu(), u64::cast_from(rhs), stream
+                            &*lhs.ciphertext.on_gpu(), u64::cast_from(rhs), stream
                         )
                     });
                     RadixCiphertext::Cuda(inner_result)
@@ -808,7 +808,7 @@ generic_integer_impl_scalar_operation!(
                 InternalServerKey::Cuda(cuda_key) => {
                     let inner_result = with_thread_local_cuda_stream(|stream| {
                         cuda_key.key.scalar_rotate_right(
-                            &lhs.ciphertext.on_gpu(), u64::cast_from(rhs), stream
+                            &*lhs.ciphertext.on_gpu(), u64::cast_from(rhs), stream
                         )
                     });
                     RadixCiphertext::Cuda(inner_result)
